pyproject.toml
[tool.poetry]
name = "web_bot"
version = "0.1.0"
description = ""
authors = ["Kamran <kamrankek48@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
aiohttp = "3.9"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
django-extensions = "^3.2.3"
python-ulid = "^3.0.0"
celery = "^5.5.3"
redis = "^6.2.0"
gevent = "^25.5.1"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.1"
django-redis = "^6.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
